Refactored get_trexolists_targets() to enable grouping by same object (using RA and Dec). Moved rank_planets(), solve_period_sma(), solve_rp_rs(), solve_a_rs() complete_entry(), and as_str() from  to catalog_utils.py. Fixed typo in fetch_trexolist() and load_trexolist_table() names (functions were names trexolits).
