+++ gen_tso/catalogs/catalog_utils.py
@@ -11,15 +11,22 @@ __all__ = [
     'is_letter',
     'is_candidate',
     'get_letter',
-    'to_float',
     'select_alias',
     'invert_aliases',
+    'rank_planets',
+    'solve_period_sma',
+    'solve_rp_rs',
+    'solve_a_rs',
+    'complete_entry',
+    'to_float',
+    'as_str',
 ]
 
 import re
 
 from astropy.io import ascii
 import numpy as np
+import pyratbay.constants as pc
 
 from ..utils import ROOT
 
@@ -71,9 +78,16 @@ def normalize_name(target):
     return name
 
 
-def get_trexolists_targets(trexo_file=None, extract='Target'):
+def get_trexolists_targets(grouped=False, trexo_file=None, extract='Target'):
     """
     Get the target names from the trexolists.csv file.
+
+    Parameters
+    ----------
+    grouped: Bool
+        If False, return a 1D list of names.
+        If True, return a nested list of lists, with each item the
+        set of names for a same object.
     """
     if trexo_file is None:
         trexo_file = f'{ROOT}data/trexolists.csv'
@@ -84,12 +98,20 @@ def get_trexolists_targets(trexo_file=None, extract='Target'):
     )
 
     targets = trexolist_data['Target'].data
+    norm_targets = [
+        normalize_name(target)
+        for target in targets
+    ]
+    if not grouped:
+        return np.unique(norm_targets)
+
+    # Use RA and dec to detect aliases for a same object
     ra = trexolist_data['R.A. 2000'].data
     dec = trexolist_data['Dec. 2000'].data
     truncated_ra = [r[0:7] for r in ra]
     truncated_dec = [d[0:6] for d in dec]
 
-    # Detect aliases for a same object
+    target_sets = []
     ntargets = len(targets)
     taken = np.zeros(ntargets, bool)
     for i in range(ntargets):
@@ -98,18 +120,14 @@ def get_trexolists_targets(trexo_file=None, extract='Target'):
         ra = truncated_ra[i]
         dec = truncated_dec[i]
         taken[i] = True
-        hosts = [targets[i]]
+        hosts = [norm_targets[i]]
         for j in range(i,ntargets):
             if truncated_ra[j]==ra and truncated_dec[j]==dec and not taken[j]:
-                hosts.append(targets[j])
+                hosts.append(norm_targets[j])
                 taken[j] = True
         # print(f'{ra}  {dec}   {np.unique(hosts)}')
-
-    norm_targets = [
-        normalize_name(target)
-        for target in targets
-    ]
-    return np.unique(norm_targets)
+        target_sets.append(list(np.unique(hosts)))
+    return target_sets
 
 
 def is_letter(name):
@@ -169,6 +187,76 @@ def invert_aliases(aliases):
     return aka
 
 
+def rank_planets(entries):
+    """
+    Rank entries with the most data
+    """
+    points = [
+        (
+            (entry['st_teff'] is None) +
+            (entry['st_logg'] is None) +
+            (entry['st_met'] is None) +
+            (entry['pl_trandur'] is None) +
+            (entry['pl_rade'] is None) +
+            (entry['pl_orbsmax'] is None and entry['pl_ratdor'] is None) +
+            (entry['st_rad'] is None and entry['pl_ratror'] is None)
+        )
+        for entry in entries
+    ]
+    rank = np.argsort(np.array(points))
+    return rank
+
+
+def solve_period_sma(period, sma, mstar):
+    """
+    Solve period-sma-mstar system values.
+    """
+    if mstar is None or mstar == 0:
+        return period, sma
+    if period is None and sma is not None:
+        period = (
+            2.0*np.pi * np.sqrt((sma*pc.au)**3.0/pc.G/(mstar*pc.msun)) / pc.day
+        )
+    elif sma is None and period is not None:
+        sma = (
+            ((period*pc.day/(2.0*np.pi))**2.0*pc.G*mstar*pc.msun)**(1/3) / pc.au
+        )
+    return period, sma
+
+
+def solve_rp_rs(rp, rs, rprs):
+    if rp is None and rs is not None and rprs is not None:
+        rp = rprs * (rs*pc.rsun) / pc.rearth
+    if rs is None and rp is not None and rprs is not None:
+        rs = rp*pc.rearth / rprs / pc.rsun
+    if rprs is None and rp is not None and rs is not None:
+        rprs = rp*pc.rearth / (rs*pc.rsun)
+    return rp, rs, rprs
+
+
+def solve_a_rs(a, rs, ars):
+    if a is None and rs is not None and ars is not None:
+        a = ars * (rs*pc.rsun) / pc.au
+    if rs is None and a is not None and ars is not None:
+        rs = a*pc.au / ars / pc.rsun
+    if ars is None and a is not None and rs is not None:
+        ars = a*pc.au / (rs*pc.rsun)
+    return a, rs, ars
+
+
+def complete_entry(entry):
+    entry['pl_rade'], entry['st_rad'], entry['pl_ratror'] = solve_rp_rs(
+        entry['pl_rade'], entry['st_rad'], entry['pl_ratror'],
+    )
+    entry['pl_orbsmax'], entry['st_rad'], entry['pl_ratdor'] = solve_a_rs(
+        entry['pl_orbsmax'], entry['st_rad'], entry['pl_ratdor'],
+    )
+    entry['pl_orbper'], entry['pl_orbsmax'] = solve_period_sma(
+        entry['pl_orbper'], entry['pl_orbsmax'], entry['st_mass']
+    )
+    return entry
+
+
 def to_float(value):
     """
     Cast string to None or float type.
@@ -177,4 +265,12 @@ def to_float(value):
         return None
     return float(value)
 
+
+def as_str(val, fmt):
+    """
+    Format as string
+    """
+    if val is None:
+        return 'None'
+    return f'{val:{fmt}}'
